+++ config/watdafudge_c/client/files/__init__.py
@@ -1,0 +1,7 @@
+from .. import Interactor
+
+class App(Interactor):
+    
+    @classmethod
+    def LoadExtend(cls, context_type, cfg):
+        return cfg