+++ src/watdafudge/client/files/__init__.py
@@ -1,0 +1,16 @@
+from .. import Interactor, Factory
+
+
+class Factory(Factory):
+    """
+    """
+    pass
+
+class App(Interactor):
+    """
+    """
+    FactoryType = Factory
+    name = 'files'
+
+
+