+++ tests/__init__.py
@@ -11,9 +11,23 @@ src_dir = path.join(
     'src'
 )
 
+conf_dir = path.join(
+    path.dirname(test_dir),
+    'config'
+)
+
 sys.path.insert(
     1,
     src_dir
 )
 
+sys.path.insert(
+    2,
+    conf_dir
+)
+
 import watdafudge
+import watdafudge_c
+
+import pytest
+watdafudge.pytest = pytest