+++ tests/test_1000_unit/test_1000_lib/test_1000_lang/test_1010_factory.py
@@ -1,0 +1,12 @@
+from watdafudge import pytest
+
+def test_1000_factory():
+
+    from watdafudge.lib.lang.factory import Factory
+
+    config = {
+        'root': 1
+    }
+    f = Factory.Default(config)
+
+    assert f.config is config