+++ tests/test_3000_validation/test_1000_wp/test_case_1000.py
@@ -1,3 +1,12 @@
-def test_init():
+from watdafudge import pytest
+
+def test_0000_init():
     import watdafudge
-    assert watdafudge.__version__ == '0.0.1a'
+    assert watdafudge.__version__
+
+def test_1000_use_case():
+    """First TDD test case
+    """
+    from watdafudge.client.files import App
+    a = App.Load()
+