+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## 0.0.2a - 2018-08-12
+### Added
+- Base libraries:
+    - watdafudge.lib.lang
+    - watdafudge.lib.client
+- Unit tests for base libraries
+- Associated config module for client framework
+
 ## 0.0.1 - 2018-08-12
 ### Added
 - Initial project creation.